create server module
