+++ src/walletSdk/Auth/index.ts
@@ -1,10 +1,5 @@
 import { AxiosInstance } from "axios";
-import StellarSdk, {
-  Transaction,
-  TransactionBuilder,
-  Keypair,
-  StellarToml,
-} from "@stellar/stellar-sdk";
+import StellarSdk, { Transaction } from "@stellar/stellar-sdk";
 import { decode } from "jws";
 
 import { Config } from "walletSdk";
@@ -15,8 +10,6 @@ import {
   InvalidTokenError,
   MissingTokenError,
   ExpiredTokenError,
-  ChallengeTxnIncorrectSequenceError,
-  ChallengeTxnInvalidSignatureError,
 } from "../Exceptions";
 import {
   AuthenticateParams,
@@ -24,10 +17,7 @@ import {
   ChallengeParams,
   ChallengeResponse,
   SignParams,
-  SignChallengeTxnParams,
-  SignChallengeTxnResponse,
 } from "../Types";
-import { parseToml } from "../Utils";
 
 export { WalletSigner, DefaultSigner } from "./WalletSigner";
 
@@ -174,46 +164,4 @@ const validateToken = (token: string) => {
   if (parsedToken.expiresAt < Math.floor(Date.now() / 1000)) {
     throw new ExpiredTokenError(parsedToken.expiresAt);
   }
-};
-
-/**
- * Helper method for signing a SEP-10 challenge transaction if valid.
- * @param {SignChallengeTxnParams} params - The Authentication params.
- * @param {AccountKeypair} params.accountKp - Keypair for the Stellar account signing the transaction.
- * @param {string} [params.challengeTx] - The challenge transaction given by an anchor for authentication.
- * @param {string} [params.networkPassphrase] - The network passphrase for the network authenticating on.
- * @param {string} [params.anchorDomain] - Domain hosting stellar.toml file containing `SIGNING_KEY`.
- * @returns {Promise<SignChallengeTxnResponse>} The signed transaction.
- */
-export const signChallengeTransaction = async ({
-  accountKp,
-  challengeTx,
-  networkPassphrase,
-  anchorDomain,
-}: SignChallengeTxnParams): Promise<SignChallengeTxnResponse> => {
-  const tx = TransactionBuilder.fromXDR(
-    challengeTx,
-    networkPassphrase,
-  ) as Transaction;
-
-  if (parseInt(tx.sequence) !== 0) {
-    throw new ChallengeTxnIncorrectSequenceError();
-  }
-
-  const tomlResp = await StellarToml.Resolver.resolve(anchorDomain);
-  const parsedToml = parseToml(tomlResp);
-  const anchorKp = Keypair.fromPublicKey(parsedToml.signingKey);
-
-  const isValid =
-    tx.signatures.length &&
-    anchorKp.verify(tx.hash(), tx.signatures[0].signature());
-  if (!isValid) {
-    throw new ChallengeTxnInvalidSignatureError();
-  }
-
-  accountKp.sign(tx);
-  return {
-    transaction: tx.toXDR(),
-    networkPassphrase,
-  };
 };
